requirements.txt
# Install mcp from GitHub directly since it's not available on PyPI
git+https://github.com/modelcontextprotocol/python-sdk.git
supabase==2.4.5
python-dotenv==1.0.0
Flask==2.3.2